requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
pyyaml==6.0.2
yfinance==0.2.52
httpx==0.27.0
python-dateutil==2.9.0.post0
pandas==2.2.2
numpy==2.0.1
